Backend/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
crayons==0.2.0
cryptography==2.7
Django==2.2.1
django-jet==1.0.8
django-oauth-toolkit==1.2.0
djangorestframework==3.9.4
idna==2.8
oauthlib==3.0.1
psycopg2==2.8.2
pycparser==2.19
pyOpenSSL==19.0.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
selenium==3.141.0
shortuuid==0.5.0
simple-salesforce==0.74.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
webdriver-manager==1.7
